More adjustments to tests. Include test for changing ids.
